pyproject.toml
[project]
name = "debt-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.0",
    "livekit-agents[openai,deepgram,cartesia,silero,turn_detector]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv~=1.0",
]
